requirements.txt
docopt
PyYAML
